refactor: reorganize CPU operation tests
Having all CPU operations in one file is cumbersome to manage because the file is quite large. This change reorganizes all tests into separate files that match the categorization of operations defined in the CPU class.

1. Adds operation_tests directory to cpu tests.
2. Separates CPU operation tests into separate files based on the type of operation.
